requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.28
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp-socks==0.8.4
fastapi==0.109.2
uvicorn==0.27.1
Jinja2==3.1.3
yookassa==3.10.0
asyncpg==0.29.0
redis==5.0.3
python-multipart==0.0.9
fpdf2==2.7.9
pydantic>=2.4.1,<2.6
google-genai
matplotlib
alembic==1.13.1
yookassa
